Forces the "correct" url for pages #239
Since lagrummet, for some reason don't care about most of the url, it's
possible to any url for a page as long as it ends with the pages
permalink.

This redirects the user to the correct path, based on where in the
site-tree the page lives.

there is a kindof exception for "home" as well.