Show versions of pages distinct #364
This fix is not perfect, since it checks the permalink. This is kind of
a edge case anyway.

The removed sort is to make it pick up the master revision, so it works
like pageController/show.

Also just picks up the _really_ published pages with the db query.
checks are also done later in like the view I think :cry:

I guess the ppl creating this site didn't know of DRY..